Restructure compaction status store packages
