+++ java/compaction-status-store/src/main/java/sleeper/compaction/status/store/CompactionStatusStoreException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.status;
+package sleeper.compaction.status.store;
 
 public class CompactionStatusStoreException extends RuntimeException {
