+++ java/compaction-status-store/src/test/java/sleeper/compaction/status/store/job/DynamoDBCompactionJobStatusStoreTimeToLiveIT.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package sleeper.compaction.status.job;
+package sleeper.compaction.status.store.job;
 
 import org.junit.Test;
 import sleeper.compaction.job.CompactionJob;
-import sleeper.compaction.status.testutils.DynamoDBCompactionJobStatusStoreTestBase;
+import sleeper.compaction.status.store.testutils.DynamoDBCompactionJobStatusStoreTestBase;
 import sleeper.core.partition.Partition;
 import sleeper.core.record.process.RecordsProcessed;
 import sleeper.core.record.process.RecordsProcessedSummary;
