+++ java/compaction-status-store/src/test/java/sleeper/compaction/status/store/job/QueryCompactionJobStatusUnfinishedIT.java
@@ -13,25 +13,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.status.job;
+package sleeper.compaction.status.store.job;
 
 import org.junit.Test;
 import sleeper.compaction.job.CompactionJob;
 import sleeper.compaction.job.TestCompactionJobStatus;
-import sleeper.compaction.status.testutils.DynamoDBCompactionJobStatusStoreTestBase;
+import sleeper.compaction.job.status.CompactionJobCreatedStatus;
+import sleeper.compaction.job.status.CompactionJobStatus;
+import sleeper.compaction.status.store.testutils.DynamoDBCompactionJobStatusStoreTestBase;
 import sleeper.core.partition.Partition;
+import sleeper.core.record.process.status.ProcessFinishedStatus;
+import sleeper.core.record.process.status.ProcessRun;
+import sleeper.core.record.process.status.ProcessStartedStatus;
 import sleeper.statestore.FileInfoFactory;
 
-import java.time.Instant;
-import java.time.Period;
+import java.util.Arrays;
 import java.util.Collections;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class QueryCompactionJobStatusByPeriodIT extends DynamoDBCompactionJobStatusStoreTestBase {
+public class QueryCompactionJobStatusUnfinishedIT extends DynamoDBCompactionJobStatusStoreTestBase {
 
     @Test
-    public void shouldReturnCompactionJobsInPeriod() {
+    public void shouldReturnUnfinishedCompactionJobs() {
         // Given
         Partition partition = singlePartition();
         FileInfoFactory fileFactory = fileFactory(partition);
@@ -47,9 +51,7 @@ public class QueryCompactionJobStatusByPeriodIT extends DynamoDBCompactionJobSta
         store.jobCreated(job2);
 
         // Then
-        Instant epochStart = Instant.ofEpochMilli(0);
-        Instant farFuture = epochStart.plus(Period.ofDays(999999999));
-        assertThat(store.getJobsInTimePeriod(tableName, epochStart, farFuture))
+        assertThat(store.getUnfinishedJobs(tableName))
                 .usingRecursiveFieldByFieldElementComparator(IGNORE_UPDATE_TIMES)
                 .containsExactly(
                         TestCompactionJobStatus.created(job2, ignoredUpdateTime()),
@@ -57,21 +59,27 @@ public class QueryCompactionJobStatusByPeriodIT extends DynamoDBCompactionJobSta
     }
 
     @Test
-    public void shouldExcludeCompactionJobOutsidePeriod() {
+    public void shouldExcludeFinishedCompactionJob() {
         // Given
         Partition partition = singlePartition();
         FileInfoFactory fileFactory = fileFactory(partition);
-        CompactionJob job = jobFactory.createCompactionJob(
-                Collections.singletonList(fileFactory.leafFile(123L, "a", "z")),
+        CompactionJob job1 = jobFactory.createCompactionJob(
+                Collections.singletonList(fileFactory.leafFile("file1", 123L, "a", "c")),
+                partition.getId());
+        CompactionJob job2 = jobFactory.createCompactionJob(
+                Collections.singletonList(fileFactory.leafFile("file2", 456L, "d", "f")),
                 partition.getId());
 
         // When
-        store.jobCreated(job);
+        store.jobCreated(job1);
+        store.jobCreated(job2);
+        store.jobStarted(job2, defaultStartTime(), DEFAULT_TASK_ID);
+        store.jobFinished(job2, defaultSummary(), DEFAULT_TASK_ID);
 
         // Then
-        Instant periodStart = Instant.now().plus(Period.ofDays(1));
-        Instant periodEnd = periodStart.plus(Period.ofDays(1));
-        assertThat(store.getJobsInTimePeriod(tableName, periodStart, periodEnd)).isEmpty();
+        assertThat(store.getUnfinishedJobs(tableName))
+                .usingRecursiveFieldByFieldElementComparator(IGNORE_UPDATE_TIMES)
+                .containsExactly(TestCompactionJobStatus.created(job1, ignoredUpdateTime()));
     }
 
     @Test
@@ -91,58 +99,38 @@ public class QueryCompactionJobStatusByPeriodIT extends DynamoDBCompactionJobSta
         store.jobCreated(job2);
 
         // Then
-        Instant epochStart = Instant.ofEpochMilli(0);
-        Instant farFuture = epochStart.plus(Period.ofDays(999999999));
-        assertThat(store.getJobsInTimePeriod(tableName, epochStart, farFuture))
+        assertThat(store.getUnfinishedJobs(tableName))
                 .usingRecursiveFieldByFieldElementComparator(IGNORE_UPDATE_TIMES)
                 .containsExactly(TestCompactionJobStatus.created(job1, ignoredUpdateTime()));
     }
 
     @Test
-    public void shouldIncludeFinishedStatusUpdateOutsidePeriod() throws Exception {
+    public void shouldIncludeUnfinishedCompactionJobWithOneFinishedRun() {
         // Given
         Partition partition = singlePartition();
         FileInfoFactory fileFactory = fileFactory(partition);
         CompactionJob job = jobFactory.createCompactionJob(
-                Collections.singletonList(fileFactory.leafFile(123L, "a", "z")),
-                partition.getId());
-
-        // When
-        Instant periodStart = Instant.now().minus(Period.ofDays(1));
-        store.jobCreated(job);
-        store.jobStarted(job, defaultStartTime(), DEFAULT_TASK_ID);
-        Thread.sleep(1);
-        Instant periodEnd = Instant.now();
-        Thread.sleep(1);
-        store.jobFinished(job, defaultSummary(), DEFAULT_TASK_ID);
-
-        // Then
-        assertThat(store.getJobsInTimePeriod(tableName, periodStart, periodEnd))
-                .usingRecursiveFieldByFieldElementComparator(IGNORE_UPDATE_TIMES)
-                .containsExactly(finishedStatusWithDefaults(job));
-    }
-
-    @Test
-    public void shouldIncludeJobCreatedOutsidePeriod() throws Exception {
-        // Given
-        Partition partition = singlePartition();
-        FileInfoFactory fileFactory = fileFactory(partition);
-        CompactionJob job = jobFactory.createCompactionJob(
-                Collections.singletonList(fileFactory.leafFile(123L, "a", "z")),
+                Collections.singletonList(fileFactory.leafFile("file1", 123L, "a", "c")),
                 partition.getId());
 
         // When
         store.jobCreated(job);
-        Thread.sleep(1);
-        Instant periodStart = Instant.now();
-        Thread.sleep(1);
+        store.jobCreated(job);
         store.jobStarted(job, defaultStartTime(), DEFAULT_TASK_ID);
         store.jobFinished(job, defaultSummary(), DEFAULT_TASK_ID);
-        Instant periodEnd = periodStart.plus(Period.ofDays(1));
+        store.jobStarted(job, defaultStartTime(), DEFAULT_TASK_ID);
 
         // Then
-        assertThat(store.getJobsInTimePeriod(tableName, periodStart, periodEnd))
+        assertThat(store.getUnfinishedJobs(tableName))
                 .usingRecursiveFieldByFieldElementComparator(IGNORE_UPDATE_TIMES)
-                .containsExactly(finishedStatusWithDefaults(job));
+                .containsExactly(CompactionJobStatus.builder().jobId(job.getId())
+                        .createdStatus(CompactionJobCreatedStatus.from(job, ignoredUpdateTime()))
+                        .jobRunsLatestFirst(Arrays.asList(
+                                ProcessRun.started(DEFAULT_TASK_ID,
+                                        ProcessStartedStatus.updateAndStartTime(ignoredUpdateTime(), defaultStartTime())),
+                                ProcessRun.finished(DEFAULT_TASK_ID,
+                                        ProcessStartedStatus.updateAndStartTime(ignoredUpdateTime(), defaultStartTime()),
+                                        ProcessFinishedStatus.updateTimeAndSummary(ignoredUpdateTime(), defaultSummary()))
+                        )).build());
     }
 }
