+++ java/compaction-status-store/src/test/java/sleeper/compaction/status/store/task/QueryAllCompactionTasksIT.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.status.task;
+package sleeper.compaction.status.store.task;
 
 import org.junit.Test;
-import sleeper.compaction.status.testutils.DynamoDBCompactionTaskStatusStoreTestBase;
+import sleeper.compaction.status.store.testutils.DynamoDBCompactionTaskStatusStoreTestBase;
 import sleeper.compaction.task.CompactionTaskStatus;
 
 import java.time.Instant;
